refactor(client): move components
